refactor: delete eth-log-handler
